requirements.txt
coloredlogs==1.0.1
decorator==4.0.2
humanfriendly==1.33
ipython==4.0.0
ipython-genutils==0.1.0
path.py==7.6.1
pexpect==3.3
pickleshare==0.5
python-telegram-bot==2.7
requests==2.7.0
simplegeneric==0.8.1
traitlets==4.0.0
wheel==0.24.0
